refactor: gql queries
